# bpo-39200: Correct the error message for range() empty constructor