--- a/Lib/test/test_range.py
+++ b/Lib/test/test_range.py
@@ -91,6 +91,13 @@ class RangeTest(unittest.TestCase):
         r = range(-sys.maxsize, sys.maxsize, 2)
         self.assertEqual(len(r), sys.maxsize)
 
+    def test_range_constructor_error_messages(self):
+        with self.assertRaisesRegex(TypeError, "range expected at least 1 argument, got 0"):
+            range()
+
+        with self.assertRaisesRegex(TypeError, "range expected at most 3 arguments, got 6"):
+            range(1, 2, 3, 4, 5, 6)
+
     def test_large_operands(self):
         x = range(10**20, 10**20+10, 3)
         self.assertEqual(len(x), 4)