--- a/Lib/test/test_range.py
+++ b/Lib/test/test_range.py
@@ -92,10 +92,16 @@ class RangeTest(unittest.TestCase):
         self.assertEqual(len(r), sys.maxsize)
 
     def test_range_constructor_error_messages(self):
-        with self.assertRaisesRegex(TypeError, "range expected at least 1 argument, got 0"):
+        with self.assertRaisesRegex(
+                TypeError,
+                "range expected at least 1 argument, got 0"
+        ):
             range()
 
-        with self.assertRaisesRegex(TypeError, "range expected at most 3 arguments, got 6"):
+        with self.assertRaisesRegex(
+                TypeError,
+                "range expected at most 3 arguments, got 6"
+        ):
             range(1, 2, 3, 4, 5, 6)
 
     def test_large_operands(self):