<!-- issue-number: [bpo-39200](https://bugs.python.org/issue39200) -->
https://bugs.python.org/issue39200
<!-- /issue-number -->