@serhiy-storchaka 2020-01-03
Other option - just use `PyArg_UnpackTuple(args, "range", 1, 3, &start, &stop, &step)` and check if `stop == NULL`.