@pablogsal 2020-01-03
> Other option - just use `PyArg_UnpackTuple(args, "range", 1, 3, &start, &stop, &step)` and check if `stop == NULL`.

I have implemented the switch statement as we would need to do the dance with `step` and `start` anyway and I find the switch statement a bit more elegant.