@pablogsal 2020-01-03
> Would be nice to add tests.

Done in 146591c