@pablogsal 2020-01-03
Btw, as a bonus the constructor of `range` is anecdotally faster:

# master
 ./python.exe -m pyperf timeit "range(1,3,4)"
.....................
Mean +- std dev: 893 ns +- 44 ns

# patched

 ./python.exe -m pyperf timeit "range(1,3,4)"
.....................
Mean +- std dev: 822 ns +- 14 ns